Changed uint32 to int64

+++ eblock.go
@@ -10,18 +10,18 @@ import (
 
 type EBlock struct {
 	Header struct {
-		BlockSequenceNumber uint32 `json:"blocksequencenumber"`
+		BlockSequenceNumber int64  `json:"blocksequencenumber"`
 		ChainID             string `json:"chainid"`
 		PrevKeyMR           string `json:"prevkeymr"`
-		Timestamp           uint32 `json:"timestamp"`
-		DBHeight            uint32 `json:"dbheight"`
+		Timestamp           int64  `json:"timestamp"`
+		DBHeight            int64  `json:"dbheight"`
 	} `json:"header"`
 	EntryList []EBEntry `json:"entrylist"`
 }
 
 type EBEntry struct {
 	EntryHash string `json:"entryhash"`
-	Timestamp uint32 `json:"timestamp"`
+	Timestamp int64  `json:"timestamp"`
 }
 
 func (e *EBlock) String() string {
